#53 moving types for digital signature and code clean up profile
