+++ cas-dotnet-sdk/DigitalSignature/Types/SHAED25519DalekDigitialSignatureResult.cs
@@ -1,0 +1,8 @@
+﻿namespace CasDotnetSdk.DigitalSignature.Types
+{
+    public class SHAED25519DalekDigitialSignatureResult
+    {
+        public byte[] PublicKey { get; set; }
+        public byte[] Signature { get; set; }
+    }
+}
