+++ cas-dotnet-sdk/DigitalSignature/Types/SHARSADigitalSignatureResult.cs
@@ -1,0 +1,9 @@
+﻿namespace CasDotnetSdk.DigitalSignature.Types
+{
+    public class SHARSADigitalSignatureResult
+    {
+        public string PublicKey { get; set; }
+        public string PrivateKey { get; set; }
+        public byte[] Signature { get; set; }
+    }
+}
