+++ cas-dotnet-sdk/Hashers/Blake2Wrapper.cs
@@ -66,7 +66,7 @@ namespace CasDotnetSdk.Hashers
                 throw new Exception("You must provide data to compare to verify with Blake 2 512");
             }
 
-            DateTime start = DateTime.UtcNow; 
+            DateTime start = DateTime.UtcNow;
             if (this._platform == OSPlatform.Linux)
             {
                 bool result = Blake2LinuxWrapper.blake2_512_bytes_verify(hashedData, hashedData.Length, toCompare, toCompare.Length);
