+++ cas-dotnet-sdk/Http/BenchmarkSender.cs
@@ -3,7 +3,6 @@ using CASHelpers.Types.HttpResponses.BenchmarkAPI;
 using System;
 using System.Net.Http;
 using System.Net.Http.Json;
-using System.Text.Json;
 using System.Threading.Tasks;
 
 namespace CasDotnetSdk.Http
